+++ cmd/ssql/main.go
@@ -70,6 +70,7 @@ func buildRootCommand() *cf.Command {
 	cmd = commands.RegisterCount(cmd)
 	cmd = commands.RegisterDescribe(cmd)
 	cmd = commands.RegisterUnpivot(cmd)
+	cmd = commands.RegisterFill(cmd)
 	cmd = commands.RegisterWhere(cmd)
 	cmd = commands.RegisterUpdate(cmd)
 	cmd = commands.RegisterCast(cmd)